Make pub alt-src push more smoothy and retryable [DELIVERY-4334]
This patch contains two changes:
  1. Previously, if the repo is at a certain state but not 'PUSHED' or
'STAGED', it will raise error and ask user to re-push with option
--restage. Now alt-src just clears the directory and restarts, cut off
the manual work.
  2. There was a bug if the state is 'STAGED' -- it doesn't checkout to
the target branch and then do git fetch directly, this causes git error,
make the task non-retryable. Now, if the state is 'STAGED', it checkouts
out the wanted branch first.

Change-Id: Id63368023995b14f7fa264eebe21c66b66987548