Refactor test cases to get coverage report for alt-src
The test cases for alt-src used to work by exceuting the alt-src
command as a separate process intead of importing it. That makes
it impossible to get the coverage report to see if the changed code
is tested or not. Now test the code by importing and calling the main
function.

Change-Id: Icf4da351440d1547c86b7a0b1fdf08b5b2f10766