Alt-src push should not fail if sources files are already pushed [DELIVERY-4321]
If the source files were already pushed to git repository, the task
used to raise an error and fail the push. This leads to confusion to
end users. And actually, if that happens, it means the tag isn't in
the remote repository.
So, instead of failing the push, now it continues pushing the tags.

Change-Id: I92c315f695109c12f5e4e3d5aa71286f07d6a222