Fix get_state to allow make_workdir to handle incomplete staging [DELIVERY-5559]
Previously, alt-src would crash if an SRPM had no state file due to
incomplete staging because the get_state method would attempt to open
the file before checking that it exists. This change fixes that issue
by doing just that, returning none if the file isn't found so that
the make_workdir method can handle this case.

Change-Id: I5c35cedc08911ecd2e0e1dc3417006beba970c26