Add coverage configuration file to clean up coverage report
Previously the coverage report generated for this project contained
a large number of uninteresting files from tests and the virtual
environment, cluttering the report and scewing coverage total. This
was due to the fact that the source files are not in their own
directory, meaning we must select the whole project for coverage.
By adding this .coveragerc file, we may now omit any python
directories or files that are not of interest.

Change-Id: I50e98c569ab7c3ce17973ac5c1ac0c995bda172f