Prepare rh-altsrc-tools for python3 CI [DELIVERY-5727]
Previously, this project was not able to be built or tested using python3
CI due to incompatibilities in tests and source code. This change fixes
that by updating some tests, adding pytest.xfail to test which attempted
to use incompatible code. Test requirements and the pytest.ini file were
also updated.

Change-Id: I86fbbebc716c8867ea73806bdf7cc396cf1a3449