Fix init remote repo by rsync [DELIVERY-5678]
Fix initialization of remote repo that is done by rsync.
There was redundant config option that was used for
determinantion of rsync destination of remote repo.
Since we have different destinations for srpms and module
sources and old rsync config used only one path for both,
it is required to use correct remote destination for both
sources.

Change-Id: Ib1770300e2eaf36390ce5f8150cafe684a1b2d08