alt-src uses pagure api to create repo if applicable [DELIVERY-6475]
CentOS source repo will be migrated from gitblit to
pagure. So alt-src uses pagure init new project api
to create a repo. It checks whether api is specified
in the config as a check if the destination is gitblit
or pagure.

The api requires a key for authorization which is
saved in a file and path to that file is specified
in alt-src.conf. alt-src.conf has the init new project
api as well.

Change-Id: Ic5e50b7885fe2221897df525a2582b2149efbdba