alt-src commits empty while creating new orphan branch during checkout [DELIVERY-6689]
alt-src creates an orphan branch if the required branch
doesn't exist in the repo. Older versions of git requires
HEAD to point to a branch while creating an orphan branch.
This is usally pointing to master. However, there's no master
available for pagure repos. HEAD in that case doesn't point to
a branch, unless the repo is initalized for the first time.
Empty commit inits the master branch with the HEAD pointing to
it. Checking out an orphan branch after that works usual.
Also, added the test case to verfiy new branch creation
where there's no master and HEAD is empty.

Change-Id: I891f22389e8c63bf079b637992a7ecf78f7737db