alt-src has separate local git/stage paths for rpms and modules [DELIVERY-7065]
Alt-src script used to clone git repo to <gitdir>/ if it's a srpm
package or <gitdir>/modules/ if it's modules. This is unclean and
hard to maintain. Now change it to <gitdir>/rpms/ if it's srpm,
make it consistent with modules.

For staging dir, alt-src now stages modules content under <stagedir>/
modules and srpms under <stagedir>/rpms.

Change-Id: Ic2e4bb39235f6a8f9dbc836958f58835657fa4d5