Remove local tag if already exists [DELIVERY-7359]
This commit adds support for scenario when one alt-src push
failed to finish but there was already local git tag created.

When re-push of the same push is triggered, it lists current
local tags and if there is a local tag present with the same name
that  is going to be created, delete the old one and create a new one.