Fix broken handling of lockfile during cleanup [DELIVERY-8817] (#23)
Previously, if the staging was incomplete, alt-src script would remove
the whole work directory together with lock file. It may leading to
crash since the os.rmdir doesn't support removing non-emtpty dir and
removal of lock file may give another process a chance creating a lock
file at the same path.
This patch rename the lock file to <workdir>.lock and make it a sibling
of workdir, which means while removing the workdir, the lock won't be
removed. Also the method used to remove dir is changed to shutil.rmtree.