tests: fix a function wrongly marked as fixture
This was marked as pytest.fixture, but it was never used as such,
it was simply called as a regular function.
This was deprecated, and on recent versions of pytest gives a
fatal error, see:
https://docs.pytest.org/en/latest/deprecations.html#calling-fixtures-directly

As this doesn't depend on any other fixtures or have any teardown
logic after the test completes, it doesn't need to be a fixture,
so just make it a regular function.