Support extracting RPMs with rpm2cpio
We need to access files within SRPMs. Previously, the method of
doing this was to install the SRPM while overriding _topdir.

Problem: RPM will refuse to do this if the SRPM is declared as
depending on unavailable rpmlib features. In general, this could
happen when taking an RPM produced for some RHEL version and
passing it to alt-src in an older RHEL version. Concretely it's
happening at the moment when RPMs using rich deps (RHEL8) are
handled from a RHEL6 host, giving errors such as:

  error: Missing rpmlib features for rust-1.43.1-1.module+el8.2.1+6788+cceb3c81.aarch64:
  error: 	rpmlib(RichDependencies) <= 4.12.0-1

When an RPM is not installable for this reason, it may be possible
to instead unpack it with rpm2cpio, so let's do that. The main
annoyance with this is that RPM applies certain relocation logic
when installing SRPMs, and this doesn't kick in when extracting
via rpm2cpio, so we have to reimplement the same thing ourselves.

Note that the rpm2cpio method of unpacking is only used if
"rpm -i" failed. Installing the RPM remains the preferred method
because cpio is a legacy format with a limit of 4GB.  Thus, if
using alt-src from an old version of RHEL to process RPMs from a
newer version, it may be the case that some RPMs only work with
one method, some with the other... and some not at all :(