Fix compatibility with old version of six
six.ensure_text was introduced in six 1.12.0, but six 1.9.0 seems
to be the latest version of the package shipped for RHEL6, so
make sure it's compatible with that. We introduce our own tiny
helper to replace ensure_text.

The CI was also updated to ensure that at least one configuration
tests against the appropriate version of six.