CI: use specific version of rpm-python in legacy tests
In the test configuration where we're trying to test the oldest
supported version of modules, let's request rpm-py-installer to
install an older version of rpm-python. This helps ensure
compatibility on older systems as far back as RHEL6.

(It'd be best if we could use version 4.8.0 for this, being the
version on RHEL6, but 4.10.0 seems to be the oldest where
installation via rpm-py-installer is able to succeed.)