Add option to clean lookaside cache for a package
Sources with large size and some other factors are filtered out
and moved to a lookaside path from where it's sent to remote.
These files are usually large in size and are not cleared after
the push completes.
This patch provides and option to clear the lookaside path for a
package before the content is pushed to lookaside during stage and
after the content is synced to remote.