Fix compatibility with latest PyYAML
yaml.load without Loader has been deprecated for a long time,
and the argument was finally made mandatory as of PyYAML 6.
Update all calls to always specify a Loader.

Note that we had already updated one yaml load to use BaseLoader
in 59282807 to fix a bug, but others were not changed at that time.

It is additionally required to explicitly set SafeDumper while
calling yaml.dump during debranding, otherwise introducing BaseLoader
would have the side-effect of introducing 'python/unicode' tags
everywhere as data is round-tripped. This is demonstrated by the
following experiments:

   >>> from yaml import dump, load, BaseLoader, SafeDumper

   # default: OK
   >>> dump(load(dump({'foo': 'bar'})))
   'foo: bar\n'

   # Adding BaseLoader: undesirable python-specific output
   >>> dump(load(dump({'foo': 'bar'}), Loader=BaseLoader))
   "!!python/unicode 'foo': !!python/unicode 'bar'\n"

   # Adding SafeDumper: OK
   # (note: per docs, BaseDumper is only intended for subclassing
   # and can't be used directly)
   >>> dump(load(dump({'foo': 'bar'}), Loader=BaseLoader), Dumper=SafeDumper)
   'foo: bar\n'